fix(mobile/scale-description): added pane title
