+++ src/components/FloatingScaleDescButton.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Fab, makeStyles } from "@material-ui/core";
 import { KnowitColors } from "../styles";
-import DescriptionTableMobile from "./DescriptionTableMobile";
+import DescriptionTable from "./DescriptionTable";
 
 const floatingScaleDescButtonStyleDesktop = makeStyles({
     fab: {
@@ -19,7 +19,7 @@ const floatingScaleDescButtonStyleDesktop = makeStyles({
         fontWeight: "bold",
         fontSize: "11px",
         lineHeight: "13px",
-        height: "35px"
+        height: "35px",
     },
     fabMenu: {
         position: "absolute",
@@ -31,20 +31,23 @@ const floatingScaleDescButtonStyleDesktop = makeStyles({
         borderRadius: "50px 50px 50px 50px",
         backgroundColor: KnowitColors.beige,
         width: "400px",
+        // viewport height - headerbar height - bottom margin height - other spacing
+	maxHeight: "calc(100vh - 66px - 55px - 20px)",
+        overflow: "auto",
         boxShadow:
-            "0px 3px 5px -1px rgba(0,0,0,0.2), 0px 6px 10px 0px rgba(0,0,0,0.14), 0px 1px 18px 0px rgba(0,0,0,0.12)"
+            "0px 3px 5px -1px rgba(0,0,0,0.2), 0px 6px 10px 0px rgba(0,0,0,0.14), 0px 1px 18px 0px rgba(0,0,0,0.12)",
     },
     closeButton: {
         marginTop: "15px",
         marginRight: "20px",
         "&:hover": {
-            color: KnowitColors.darkGreen
+            color: KnowitColors.darkGreen,
         },
         float: "right",
         position: "absolute",
         right: "0px",
-        top: "0px"
-    }
+        top: "0px",
+    },
 });
 
 const floatingScaleDescButtonStyleMobile = makeStyles({
@@ -62,7 +65,7 @@ const floatingScaleDescButtonStyleMobile = makeStyles({
         fontWeight: "bold",
         fontSize: "11px",
         lineHeight: "13px",
-        height: "35px"
+        height: "35px",
     },
     fabMenu: {
         position: "absolute",
@@ -71,22 +74,22 @@ const floatingScaleDescButtonStyleMobile = makeStyles({
         width: "100%",
         height: "100%",
         zIndex: 101,
-        // borderRadius: '50px 50px 0px 0px',
         backgroundColor: KnowitColors.beige,
-        color: KnowitColors.darkBrown
-        // boxShadow: '0px -2px 4px rgba(0, 0, 0, 0.25)'
+        color: KnowitColors.darkBrown,
+        boxShadow: "0px -4px 4px rgba(0, 0, 0, 0.15)",
+        borderRadius: "50px 50px 0px 0px",
     },
     closeButton: {
         marginTop: "10px",
         marginRight: "15px",
         "&:hover": {
-            color: KnowitColors.darkGreen
+            color: KnowitColors.darkGreen,
         },
         float: "right",
         position: "absolute",
         right: "0px",
-        top: "0px"
-    }
+        top: "0px",
+    },
 });
 
 type FloatingScaleDescButtonProps = {
@@ -94,7 +97,7 @@ type FloatingScaleDescButtonProps = {
 };
 
 const FloatingScaleDescButton = ({
-    isMobile
+    isMobile,
 }: FloatingScaleDescButtonProps) => {
     const style = isMobile
         ? floatingScaleDescButtonStyleMobile()
@@ -105,8 +108,9 @@ const FloatingScaleDescButton = ({
         <>
             {scaleDescOpen && (
                 <div className={style.fabMenu}>
-                    <DescriptionTableMobile
+                    <DescriptionTable
                         onClose={() => setScaleDescOpen(false)}
+                        isMobile={isMobile}
                     />
                 </div>
             )}
